--- a/packages/server/src/db/index.js
+++ b/packages/server/src/db/index.js
@@ -672,6 +672,23 @@ function validateSearchFilters(filters) {
     return errors;
 }
 
+function addCsvData(qb) {
+    qb
+        .select(knex.raw(`
+            CASE
+            WHEN grants.funding_instrument_codes = 'G' THEN 'Grant'
+            WHEN grants.funding_instrument_codes = 'CA' THEN 'Cooperative Agreement'
+            WHEN grants.funding_instrument_codes = 'PC' THEN 'Procurement Contract'
+            ELSE 'Other'
+            END as funding_type
+        `))
+        .select(knex.raw(`array_to_string(array_agg(${TABLES.eligibility_codes}.label), '|') AS eligibility`))
+        .leftJoin(
+            `${TABLES.eligibility_codes}`,
+            `${TABLES.eligibility_codes}.code`, '=', knex.raw(`ANY(string_to_array(${TABLES.grants}.eligibility_codes, ' '))`),
+        );
+}
+
 /*
    filters: {
         reviewStatuses: List[Enum['Applied', 'Not Applying', 'Interested']],
@@ -693,15 +710,15 @@ function validateSearchFilters(filters) {
     tenantId: number
     agencyId: number
 */
-async function getGrantsNew(filters, paginationParams, orderingParams, tenantId, agencyId) {
-    console.log(JSON.stringify([filters, paginationParams, orderingParams, tenantId, agencyId]));
+async function getGrantsNew(filters, paginationParams, orderingParams, tenantId, agencyId, toCsv) {
+    console.log(JSON.stringify([filters, paginationParams, orderingParams, tenantId, agencyId, toCsv]));
 
     const errors = validateSearchFilters(filters);
     if (errors.length > 0) {
         throw new Error(`Invalid filters: ${errors.join(', ')}`);
     }
 
-    const data = await knex(TABLES.grants)
+    const query = knex(TABLES.grants)
         .select([
             'grants.grant_id',
             'grants.grant_number',
@@ -771,6 +788,10 @@ async function getGrantsNew(filters, paginationParams, orderingParams, tenantId,
             'grants.funding_instrument_codes',
             'grants.bill',
         );
+    if (toCsv) {
+        query.modify(addCsvData);
+    }
+    const data = await query;
 
     const fullCount = data.length > 0 ? data[0].full_count : 0;