--- a/packages/server/src/routes/grants.js
+++ b/packages/server/src/routes/grants.js
@@ -74,6 +74,7 @@ router.get('/next', requireUser, async (req, res) => {
         await db.buildOrderingParams(req.query.ordering),
         user.tenant_id,
         user.agency_id,
+        false,
     );
 
     res.json(grants);
@@ -109,6 +110,7 @@ router.get('/exportCSVNew', requireUser, async (req, res) => {
         await db.buildOrderingParams(req.query.ordering),
         user.tenant_id,
         user.agency_id,
+        true,
     );
 
     // Generate CSV
@@ -143,7 +145,7 @@ router.get('/exportCSVNew', requireUser, async (req, res) => {
             { key: 'title', header: 'Title' },
             { key: 'viewed_by', header: 'Viewed By' },
             { key: 'interested_agencies', header: 'Interested Agencies' },
-            { key: 'opportunity_status', header: 'Status' },
+            { key: 'opportunity_status', header: 'Opportunity Status' },
             { key: 'opportunity_category', header: 'Opportunity Category' },
             { key: 'cost_sharing', header: 'Cost Sharing' },
             { key: 'award_ceiling', header: 'Award Ceiling' },
@@ -152,6 +154,10 @@ router.get('/exportCSVNew', requireUser, async (req, res) => {
             { key: 'agency_code', header: 'Agency Code' },
             { key: 'grant_id', header: 'Grant Id' },
             { key: 'url', header: 'URL' },
+            { key: 'funding_type', header: 'Funding Type' },
+            { key: 'bill', header: 'Appropriations Bill' },
+            { key: 'agency_code', header: 'Agency Code' },
+            { key: 'eligibility', header: 'Eligibility' },
         ],
     });